Fruity pebbles [3/4]
Although not required because this doesn't use any of OnePlus' code,
I want to say thanks to OnePlus for the inspiration behind this layout.

I think is among the best out there and that's why I wanted to mirror
it for our Accent picker :-)

- Added condition for DU-Tweaks preference icons to be themed a different
color for the dark theme. This was an issue with the dark theme while using
the white accent. The white preference background with white icons wasn't
working out so well.

Change-Id: I577d84b40f7515cb4638c31c2a8a5ab28051ef3e
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>